@@ -5,13 +5,24 @@
     "license": "MIT",
     "keywords": ["cakephp","cakephp bake","cakephp scaffold"],
     "require": {
-        "php": "^8.0",
-        "cakephp/cakephp": "^4.2",
-        "cakephp/bake": "^2.1"
+        "php": "^8.1",
+        "cakephp/cakephp": "^5.0",
+        "cakephp/bake": "^3.0"
+    },
+    "require-dev": {
+        "josegonzalez/dotenv": "^3.2",
+        "phpunit/phpunit": "^10.0"
     },
     "autoload": {
         "psr-4": {
             "MixerApi\\Bake\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "MixerApi\\Bake\\": "plugins/bake/src/",
+            "MixerApi\\Bake\\Test\\": "plugins/bake/tests/",
+            "MixerApi\\Bake\\Test\\App\\": "plugins/bake/tests/test_app/src/"
         }
     },
     "support": {
@@ -23,5 +34,11 @@
             "name": "MixerApi",
             "role": "Organization"
         }
-    ]
+    ],
+    "config": {
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "cakephp/plugin-installer": true
+        }
+    }
 }
