@@ -8,12 +8,10 @@ use Cake\TestSuite\Fixture\TestFixture;
 /**
  * ActorsFixture
  */
-class ActorsFixture  extends \App\Test\Fixture\ActorsFixture
+class ActorsFixture  extends TestFixture
 {
     /**
-     * Init method
-     *
-     * @return void
+     * @inheritdoc
      */
     public function init(): void
     {
